requirements.txt
yt-dlp
faster_whisper
ffmpeg-python
torch
openai
neo4j
python-dotenv
django
requests
langchain_openai
langchain_google_genai
django-tailwind
python-slugify
rouge-score
numpy